Refactor upgrade definitions into data and hooks
